"Fix" TS errors
In the `App.vue` file, a string that surely is limited to what the
function para requires has been "casted" with `as` so TS does not show
it as an error.

In the `ports.ts` an unknow error was happening, as it was unknown, a
`@ts-ignore` directive has been added above.