Implement data export
Implements two types of export, CSV data storing TUI, full name, time
and if its a new card so it can be brought into the original census; and
a LimeSurvey compatible CSV export.

The export is handled by a custom component `CSVExportButton`, which now
replaces the original button in the `App.vue` main page. The logic of
writing to the csv is handled as Rust commands, providing also a generic
way of defining more types of expports.